Eclipse warnings
git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@17412 ab1d8caa-1f67-47f1-9e81-24633a41865c

+++ sipXconfig/web/src/org/sipfoundry/sipxconfig/site/cdr/CdrReports.java
@@ -448,7 +448,7 @@ public abstract class CdrReports extends BaseComponent implements PageBeginRende
                     outgoingCalls.put(extension,
                             new CdrMinutesGraphBean(extension, (double) cdr.getDuration()));
                 } else {
-                    bean.setMinutes((bean.getMinutes() + (double) cdr.getDuration()));
+                    bean.setMinutes((bean.getMinutes() + cdr.getDuration()));
                 }
             }
         }
